+++ CHANGELOG.md
@@ -1,12 +1,30 @@
 # Changelog
 
-## 0.0.1 (UNRELEASED)
-
-- Made a nodejs test work with jaeger on ubuntu 16.04 and node.js 4
+## 0.1.0 (unreleased)
 
 ### BREAKING CHANGES
 
 ### Features
+
+- Added a secondary call to test the father-child structure in tracing
+- Added tests with dgoss
+
+### Changes
+
+- Using ubuntu 18.04
+- Using node 8.10.0
+
+### Bugfixes
+
+### Known bugs
+
+## 0.0.1
+
+### BREAKING CHANGES
+
+### Features
+
+- Made a nodejs test work with jaeger on ubuntu 16.04 and node.js 4
 
 ### Changes